requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
cassandra-driver>=3.28.0
astrapy>=0.7.0
slowapi>=0.1.8
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
pylint>=3.0.0
numpy>=1.24.0 